# ms-customer/src/main/resources/openapiswagger.yml
openapi: 3.0.3
info:
  title: Customer API
  description: API REST para gestionar clientes
  version: 1.0.0
servers:
  - url: http://localhost:8085/api/v1
    description: Servidor local de desarrollo

paths:
  /customers:
    get:
      summary: Listar todos los clientes
      operationId: getAllCustomer
      tags:
        - Customers
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'

    post:
      summary: Crear un nuevo cliente
      operationId: createCustomer
      tags:
        - Customers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerInput'
      responses:
        '201':
          description: Cliente creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  /customers/{id}:
    get:
      summary: Obtener un cliente por ID
      operationId: getCustomerById
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Cliente no encontrado

    put:
      summary: Actualizar un cliente
      operationId: updateCustomer
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerInput'
      responses:
        '200':
          description: Cliente actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Cliente no encontrado

    delete:
      summary: Eliminar un cliente
      operationId: deleteCustomer
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Cliente eliminado
        '404':
          description: Cliente no encontrado

components:
  schemas:
    Customer:
      type: object
      properties:
        idCustomer:
          type: integer
          example: 1
        firstName:
          type: string
          example: Juan
        lastName:
          type: string
          example: Pérez
        dni:
          type: string
          example: "12345678"
        email:
          type: string
          format: email
          example: juan.perez@email.com

    CustomerInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dni:
          type: string
        email:
          type: string
          format: email
      required:
        - firstName
        - lastName
        - dni
        - email